\name{sbfac-package}
\alias{sbfac-package}
\alias{sbfac}
\docType{package}
\title{
Semiparametric Bayesian Factor-Analytic Copulae for Mixed Data
}
\description{
This package provides MCMC model fitting for the semiparametric Bayesian factor-analytic copula model,
an analogue of traditional factor analysis where inference is based on the (unobserved) correlation
matrix of a Gaussian copula instead of the covariance or correlation matrix of the data itself. It handles
ordinal data that is discrete or continuous using an approximate rank-based likelihood.
}
\details{
\tabular{ll}{
Package: \tab sbfac\cr
Type: \tab Package\cr
Version: \tab 0.1.9.7\cr
Date: \tab 2011-03-17\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}

}
\author{
Jared Murray \email{jared.murray@stat.duke.edu}
}